Refactored and Moved model files to the models package
