+++ src/main/java/dev/paigewatson/layoutmaster/client/services/FreightCarService.java
@@ -1,6 +1,6 @@
 package dev.paigewatson.layoutmaster.client.services;
 
-import dev.paigewatson.layoutmaster.data.models.FreightCarDto;
+import dev.paigewatson.layoutmaster.models.data.FreightCarDto;
 
 import java.util.List;
