--- src/main/java/dev/paigewatson/layoutmaster/data/models/EntityDto.java
@@ -1,7 +1,0 @@
-package dev.paigewatson.layoutmaster.data.models;
-
-import dev.paigewatson.layoutmaster.data.NullableEntity;
-
-public interface EntityDto<T> extends NullableEntity, CastAsEntity<T>
-{
-}
