+++ src/main/java/dev/paigewatson/layoutmaster/models/NullableEntity.java
@@ -1,7 +1,6 @@
-package dev.paigewatson.layoutmaster.data;
+package dev.paigewatson.layoutmaster.models;
 
 public interface NullableEntity
 {
-
     boolean isNull();
 }
