+++ src/main/java/dev/paigewatson/layoutmaster/models/data/CarTypeDto.java
@@ -1,0 +1,5 @@
+package dev.paigewatson.layoutmaster.models.data;
+
+public interface CarTypeDto<T> extends EntityDto<T>
+{
+}
