+++ src/main/java/dev/paigewatson/layoutmaster/models/data/EntityDto.java
@@ -1,0 +1,7 @@
+package dev.paigewatson.layoutmaster.models.data;
+
+import dev.paigewatson.layoutmaster.models.NullableEntity;
+
+public interface EntityDto<T> extends NullableEntity, CastAsEntity<T>
+{
+}
