+++ src/main/java/dev/paigewatson/layoutmaster/models/rollingstock/CarType.java
@@ -1,10 +1,11 @@
 package dev.paigewatson.layoutmaster.models.rollingstock;
 
 import dev.paigewatson.layoutmaster.data.CarTypeDAL;
-import dev.paigewatson.layoutmaster.data.NullableEntity;
+import dev.paigewatson.layoutmaster.models.NullableEntity;
+import dev.paigewatson.layoutmaster.models.data.CarTypeDto;
 import dev.paigewatson.layoutmaster.models.goods.GoodsType;
 
-public interface CarType extends NullableEntity
+public interface CarType extends NullableEntity, ConvertsToDto<CarTypeDto<AARType>>
 {
     boolean canCarry(GoodsType goodsType);
 
@@ -14,5 +15,4 @@ public interface CarType extends NullableEntity
 
     CarType saveToDatabase(CarTypeDAL carTypeDAL);
 
-    CarTypeDto getDto();
 }
