--- src/main/java/dev/paigewatson/layoutmaster/models/rollingstock/CarTypeDto.java
@@ -1,7 +1,0 @@
-package dev.paigewatson.layoutmaster.models.rollingstock;
-
-import dev.paigewatson.layoutmaster.data.models.EntityDto;
-
-public interface CarTypeDto<T> extends EntityDto<T>
-{
-}
