+++ src/main/java/dev/paigewatson/layoutmaster/models/rollingstock/ConvertsToDto.java
@@ -1,0 +1,6 @@
+package dev.paigewatson.layoutmaster.models.rollingstock;
+
+public interface ConvertsToDto<T>
+{
+    T getDto();
+}
