+++ src/main/java/dev/paigewatson/layoutmaster/models/rollingstock/NullCarType.java
@@ -1,7 +1,8 @@
 package dev.paigewatson.layoutmaster.models.rollingstock;
 
 import dev.paigewatson.layoutmaster.data.CarTypeDAL;
-import dev.paigewatson.layoutmaster.data.models.NullCarTypeDto;
+import dev.paigewatson.layoutmaster.models.data.CarTypeDto;
+import dev.paigewatson.layoutmaster.models.data.NullCarTypeDto;
 import dev.paigewatson.layoutmaster.models.goods.GoodsType;
 
 public class NullCarType implements CarType
