+++ src/main/java/dev/paigewatson/layoutmaster/models/rollingstock/RollingStock.java
@@ -1,0 +1,7 @@
+package dev.paigewatson.layoutmaster.models.rollingstock;
+
+import dev.paigewatson.layoutmaster.models.NullableEntity;
+
+public interface RollingStock extends NullableEntity, ConvertsToDto<FreightCar>
+{
+}
